Create reusable CSV export hook using generics and simplify ProjectReadSetsTab
